pyproject.toml
[tool.poetry]
name = "fastapisqlalchemypytestalembic"
version = "0.1.0"
description = ""
authors = ["arletino <titkov@nxt.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
asyncpg = "^0.30.0"
alembic = "^1.13.3"
pydantic-settings = "^2.6.0"
pydantic = "^2.9.2"
fastapi = "^0.115.3"
uvicorn = "^0.32.0"
psycopg = {extras = ["binary", "pool"], version = "^3.2.3"}
greenlet = "^3.1.1"
uvloop = "^0.21.0"
#numpy = "^2.1.3"
pandas = "^2.2.3"
pympler = "^1.1"
memory-profiler = "^0.61.0"
psutil = "^6.1.0"
openpyxl = "^3.1.5"


[tool.poetry.group.dev.dependencies]
httpx = "^0.27.2"
sqlalchemy-utils = "^0.41.2"
pytest = "^8.3.3"
yarl = "^1.16.0"
mypy = "^1.13.0"
black = "^24.10.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
